refactor: Add subcommands for asg and component, under which version set and check commands are present to make the command more intuitive
