+++ cmd/k8sclusterupgradetool/component_version.go
@@ -1,0 +1,18 @@
+package k8sclusterupgradetool
+
+import (
+	"fmt"
+	"github.com/spf13/cobra"
+)
+
+var componentVersionCmd = &cobra.Command{
+	Use: "version",
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Println("K8s cluster component version operations")
+		fmt.Println("Run 'k8sclusterupgradetool component version --help' to see the available commands")
+	},
+}
+
+func init() {
+	componentCmd.AddCommand(componentVersionCmd)
+}
