+++ cmd/k8sclusterupgradetool/root.go
@@ -1,0 +1,19 @@
+package k8sclusterupgradetool
+
+import (
+	"fmt"
+	"github.com/spf13/cobra"
+	"os"
+)
+
+var RootCmd = &cobra.Command{
+	Use:   "k8sclusterupgradetool",
+	Short: "k8sclusterupgradetool",
+}
+
+func Execute() {
+	if err := RootCmd.Execute(); err != nil {
+		fmt.Println(err)
+		os.Exit(1)
+	}
+}
