--- cmd/root.go
@@ -1,19 +1,0 @@
-package cmd
-
-import (
-	"fmt"
-	"github.com/spf13/cobra"
-	"os"
-)
-
-var RootCmd = &cobra.Command{
-	Use:   "k8s-cluster-upgrade-tool",
-	Short: "k8s-cluster-upgrade-tool",
-}
-
-func Execute() {
-	if err := RootCmd.Execute(); err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
-}
