+++ e2e/component-version-check.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+set -e
+
+./k8sclusterupgradetool postUpgradeCheck kind-k8s-cluster-upgrade-tool-test-cluster
