+++ main.go
@@ -1,9 +1,9 @@
 package main
 
 import (
-	"github.com/deliveryhero/k8s-cluster-upgrade-tool/cmd"
+	"github.com/deliveryhero/k8s-cluster-upgrade-tool/cmd/k8sclusterupgradetool"
 )
 
 func main() {
-	cmd.Execute()
+	k8sclusterupgradetool.Execute()
 }
